refactor(core): split app event handler scope wiring
Extract dialog builder registration and session command subscriptions from AppEventHandlerScope into dedicated part files so lifecycle setup stays focused and maintainable while preserving existing event behavior.

Made-with: Cursor
